move useeffect to paymentprovider
